Using our own directory in the home directory
Try to use (create if not exists) the .qtt (Qtt for Windows OS)
directory in the users home directory for storing the "times"-file with
the tasks. If our own directory is not accesable or writeable fall back
to the current directory.

The currently used directroy is held in a private QDir member.