Rewrite and separate the conversion functions for seconds <-> time string
Don't use QTime anymore for the conversion because QTime only knows
time up to "23:59:59". Calculate the values for hours, minutes and seconds
and handle each value separate to build the string or value of seconds.

Move the conversion functions into its separate file: support.cpp.
This should ease the use of unit tests for the functions.

Signed-off-by: Frank Bergmann <github.qtt@frajasalo.de>