Interface-class for storage of the task data
This class defines an interface to store the data of the tasks. The functions
will be called by the GUI. A task is referenced by an id which is an unsigned
32 bit integer. The (visual) name of a task is held by the storage class.
Saving and loading use list of tasks which is implemented as a vector of task
structures. The structure is prepared for hierarchically tasks but this feature
is not used for the moment. The time value in the structure is a string at
present and should be converted to a value of seconds (unix time).

Signed-off-by: Frank Bergmann <github@frajasalo.de>