First storage class that implements the original text file format
This class saves the tasks as simple text files like the original code. One
task consists of three lines in the text file:
  1. name of the task
  2. cumulated time as string (HH:MM:SS)
  3. completed flag as string
The default file name for saving is .qtimetracker but this can be adjusted by
a function. This class does not support task periods and hierarchical tasks as
the original code. The task names will be held in a std::map in conjunction
with the id for easier access. Adding and deleting tasks change only the
internal map.