Use the new store model to save the tasks
The tree needs a new column to track the task id. This column will be hide
from the user.

As store is the StorageSimpleText class used. The Load() and Save() functions
use the IStorage-interface now. When a running task is stopped the storage
gets the chance to save the time period by calling it's Update() function.
Adding and deleting tasks invoke the corresponding functions in the store as
well.

Signed-off-by: Frank Bergmann <github@frajasalo.de>